Add feature toggle for payment
